requirements.txt
Django
djangorestframework
PyMySQL
cryptography
Pillow
requests
stripe
razorpay
python-dotenv
django-cors-headers
selenium
webdriver-manager
tqdm
mistralai<2
